+++ src/main/java/digi/alm/orm/page/Sorter.java
@@ -1,0 +1,22 @@
+package digi.alm.orm.page;
+
+import lombok.ToString;
+
+/**
+ * </p>
+ * 顺序
+ * <p>
+ *
+ * @author chenquan chenquan.dev@foxmail.com 2021/05/10 11:54
+ */
+@ToString
+public enum Sorter {
+    /**
+     * 降序
+     */
+    DESC,
+    /**
+     * 升序
+     */
+    ASC,
+}
