--- src/main/java/kd/alm/orm/core/BaseService.java
@@ -1,13 +1,0 @@
-package kd.alm.orm.core;
-
-
-/**
- * </p>
- * 基础服务层
- * <p>
- *
- * @author chenquan chenquan.dev@foxmail.com 2021/04/20 11:32
- */
-public interface BaseService<T>  extends BaseRepository<T> {
-
-}
