--- src/main/java/kd/alm/orm/core/impl/BaseServiceImpl.java
@@ -1,15 +1,0 @@
-package kd.alm.orm.core.impl;
-
-
-import kd.alm.orm.core.BaseService;
-
-/**
- * </p>
- * 基础服务层实现
- * <p>
- *
- * @author chenquan chenquan.dev@foxmail.com 2021/04/20 11:45
- */
-public class BaseServiceImpl<T> extends BaseRepositoryImpl<T> implements BaseService<T> {
-
-}
