--- src/main/java/kd/alm/orm/util/AlmEntityMetadataCacheUtils.java
@@ -1,14 +1,0 @@
-package kd.alm.orm.util;
-
-import kd.bos.entity.EntityMetadataCache;
-
-/**
- * </p>
- * 实体元数据工具类
- * <p>
- *
- * @author chenquan chenquan.dev@foxmail.com 2022/01/27 19:48
- */
-public class AlmEntityMetadataCacheUtils extends EntityMetadataCache {
-
-}
