--- src/main/java/kd/alm/orm/util/AlmSaveServiceHelper.java
@@ -1,18 +1,0 @@
-package kd.alm.orm.util;
-
-import kd.bos.logging.Log;
-import kd.bos.logging.LogFactory;
-import kd.bos.servicehelper.operation.SaveServiceHelper;
-
-/**
- * </p>
- * 数据保存工具类
- * <p>
- *
- * @author chenquan chenquan.dev@foxmail.com 2021/04/19 17:44
- */
-public class AlmSaveServiceHelper extends SaveServiceHelper {
-    private final static Log log = LogFactory.getLog(AlmSaveServiceHelper.class);
-
-
-}
